requirements.txt
flask
torch
transformers
werkzeug
pdfplumber
pdf2image
pytesseract
pillow
python-docx